check for order when deleting

+++ orders/internal/api/order_handler.go
@@ -94,12 +94,19 @@ func (h *OrderHandler) UpdateOrder(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *OrderHandler) DeleteOrder(w http.ResponseWriter, r *http.Request) {
+
 	orderID, err := uuid.Parse(mux.Vars(r)["order_id"])
 	if err != nil {
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
-
+	getOrder, err := h.orderRepository.GetOrder(orderID)
+	if getOrder == nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		w.Header().Set("Content-Type", "application/json")
+		w.Write([]byte("Order does not exist with ID " + orderID.String()))
+		return
+	}
 	err = h.orderRepository.DeleteOrder(orderID)
 	if err != nil {
 		w.WriteHeader(http.StatusInternalServerError)
